chore(refactor): files and folder structures
