+++ src/components/Navbar/Navbar.jsx
@@ -9,7 +9,7 @@ import Modal from "../Modal";
 import { FaBars } from "react-icons/fa";
 import { IoMdClose } from "react-icons/io";
 import Button from "../Button/GlobalButton/Button";
-import ClickAwayListener from "../../utils/clickAwayListener";
+import ClickAwayListener from "../../Utils/clickAwayListener";
 
 const Navbar = () => {
   const navigate = useNavigate();
