+++ src/pages/AboutUs.jsx
@@ -1,7 +1,7 @@
 import React from "react";
 import "../styles/AboutUs.css";
 // import Navbar from "../components/Navbar/Navbar";
-// import Footer from "../components/Footer";
+// import Footer from "../components/Footer/Footer";
 import abtImg from "../assets/pictures/aboutus/about-pana.png";
 
 export default function AboutUs() {
