+++ src/pages/display/ClubsPage.jsx
@@ -4,11 +4,11 @@ import { Helmet } from "react-helmet-async";
 import Loading from "../../components/Loading";
 import SingleClubEvent from "../../components/Cards/SingleClubEvent/SingleClubEvent";
 import useSWR from "swr";
-import { defaultfetcher } from "../../utils/fetcher";
-import { filter } from "../../utils/filter";
-import { showErrorToast } from "../../utils/showToast";
+import { defaultfetcher } from "../../Utils/fetcher";
+import { filter } from "../../Utils/filter";
+import { showErrorToast } from "../../Utils/Toasts";
 import Navbar from "../../components/Navbar/Navbar";
-import Footer from "../../components/Footer";
+import Footer from "../../components/Footer/Footer";
 
 const ClubsPage = () => {
   const { data: clubData, isLoading } = useSWR(
