+++ src/pages/display/EventsPage.jsx
@@ -4,11 +4,11 @@ import { Helmet } from "react-helmet-async";
 import Loading from "../../components/Loading";
 import SingleClubEvent from "../../components/Cards/SingleClubEvent/SingleClubEvent";
 import useSWR from "swr";
-import { defaultfetcher } from "../../utils/fetcher";
-import { filter } from "../../utils/filter";
+import { defaultfetcher } from "../../Utils/fetcher";
+import { filter } from "../../Utils/filter";
 // import states from "./StatesData";
-import { showErrorToast } from "../../utils/showToast";
-import Footer from "../../components/Footer";
+import { showErrorToast } from "../../Utils/Toasts";
+import Footer from "../../components/Footer/Footer";
 import Navbar from "../../components/Navbar/Navbar";
 
 const EventsPage = () => {
